// src/types/word-type.ts
export interface Phonetic {
	text: string;
	audio: string;
	sourceUrl: string;
	license: {
		name: string;
		url: string;
	};
}

export interface Meaning {
	partOfSpeech: string;
	definitions: {
		definition: string;
		synonyms?: string[];
		antonyms?: string[];
		example?: string;
	}[];
	synonyms?: string[];
	antonyms?: string[];
}

export interface WordApiResponse {
	word: string;
	phonetic: string;
	phonetics: Phonetic[];
	meanings: Meaning[];
	license: {
		name: string;
		url: string;
	};
	sourceUrls: string[];
}